% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_raw.R
\docType{data}
\name{data_raw}
\alias{data_raw}
\alias{data_processed}
\title{Retail weekly sales (Raw dataset)}
\format{
A data frame with 4,400 rows and 8 columns.
}
\source{
Cohen, M. C., Gras, P. E., Pentecoste, A., & Zhang, R. (2022).
  \emph{Demand Prediction in Retail - A Practical Guide to Leverage Data and 
  Predictive Analytics}.
  Springer Series in Supply Chain Management 14, 1–155.
}
\usage{
data(data_raw)
}
\description{
The concepts and demand prediction methods covered in the book
*Demand Prediction in Retail - A Practical Guide to Leverage Data and Predictive Analytics*
(Cohen, Gras, Pentecoste, & Zhang, 2022) are illustrated using this retail 
dataset.

This raw dataset reports the weekly sales of 44 stock-keeping units (SKUs) 
from a tech-gadget
e-commerce retailer over 100 weeks (October 2016 to September 2018), 
yielding a total of
4,400 rows. It includes the following eight columns:
\describe{
  \item{\code{week}}{Integer. Week identifier (1 to 100).}
  \item{\code{SKU}}{Character or factor. SKU identifier (44 unique SKUs).}
  \item{\code{weekly_sales}}{Numeric. Total weekly sales for the SKU in that 
  week.}
  \item{\code{featured_on_main_page}}{Logical or factor. Indicates whether 
  the SKU was featured on the main homepage (TRUE/FALSE).}
  \item{\code{color}}{Character or factor. Main color attribute of the SKU.}
  \item{\code{price}}{Numeric. Unit price of the SKU.}
  \item{\code{vendor}}{Character or factor. The supplier or vendor of the 
  SKU.}
  \item{\code{functionality}}{Character or factor. Functionality category or 
  classification.}
}

A processed version of this dataset (not included here) applies additional 
data-cleaning steps
(handling missing data, adding lagged prices, etc.). More details on the 
dataset and the
pre-processing steps can be found in the referenced book.
}
\keyword{datasets}
