% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rma.R
\docType{package}
\name{rma-package}
\alias{rma-package}
\alias{rma}
\title{Functions the final retail and marketing analytics project}
\description{
A dashboard for the final project on retail and marketing analytics
}
\details{
This is package contains the functions necessary for the final project on
retail and marketing analytics from teh Master in Business Analytics of 
Imperial College Business School
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/l-ramirez-lopez/imperial_retail_project}
}

}
